[quant][core][improvement] Made quantized conv and linear ops use reduce_range when using qnnpack backend
Summary:
Previously, the backend code "silently" ignores reduce_range=true when
using the qnnpack backend (which does not require a reduction in range).
We evaluated either 1) respecting the reduction in range to conform with
other backends (e.g., fbgemm) even when qnnpack does support the full
range and outputting a warning to let the user know that reduce_range
shoudl be set to false for qnnpack backend 2) throwing a warning and letting the user know that the
reduce_range=true setting is being ignored.

Option 1 would halve the range which could have some negative
implications to accuracy; however, we believe it is more important to
respect what the user is requesting and there have chosen to implement this option.

Fixes #68278

Test plan:

ghstack-source-id: d2442d64bbe58d5db8f78c66ee096c66ca5782af
Pull Request resolved: https://github.com/pytorch/pytorch/pull/79273